[SELS-TASK] Integrate user edit profile markup and api
