+++ backend/config/filesystems.php
@@ -38,7 +38,7 @@ return [
         'public' => [
             'driver' => 'local',
             'root' => storage_path('app/public'),
-            'url' => env('APP_URL').'/storage',
+            'url' => env('BACKEND_APP_URL') . '/storage',
             'visibility' => 'public',
         ],
