+++ frontend/src/components/UserProfile/index.tsx
@@ -1,3 +1,5 @@
 export * from './Activities';
 export * from './UserProfile';
 export * from './UserProfileEdit';
+export * from './UserProfileEditAvatar';
+export * from './UserProfileEditDetails';
